page_analyzer/app.py
import os
from flask import Flask
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)
app.config['SECRET_KEY'] = os.getenv('SECRET_KEY', 'dev-secret-key')

@app.route('/')
def index():
    return 'Hello from Page Analyzer!'

@app.route('/about')
def about():
    return 'This is an about page.'
